refactor: split orchestrate.route.ts into focused modules (Phase 4b)
Split the 2005-line orchestrate.route.ts into 7 focused modules:
- _helpers.ts: Shared schemas and utilities
- tasks.route.ts: Core task management (~380 lines)
- sync.route.ts: Head agent sync (~300 lines)
- sessions.route.ts: Provider session binding (~200 lines)
- events.route.ts: SSE real-time updates (~290 lines)
- approvals.route.ts: Human-in-the-loop approvals (~200 lines)
- learning.route.ts: Self-improving rules (~170 lines)
- index.ts: Combined router with re-exports

Target achieved: No route file exceeds 500 lines.
